+++ app/components/content/cloudinary/CloudinaryVideo.vue
@@ -1,6 +1,10 @@
 <template>
   <div class="mx-auto">
-    <CldVideoPlayer width="1920" height="1080" :src="props.videoId" />
+    <CldVideoPlayer
+      width="1920"
+      height="1080"
+      :src="props.videoId"
+    />
   </div>
 </template>